Refactoring algorithms package
