+++ src/main/java/tf/gpx/edit/algorithms/reducer/IWaypointReducer.java
@@ -23,7 +23,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.algorithms;
+package tf.gpx.edit.algorithms.reducer;
 
 import java.util.List;
 import tf.gpx.edit.helper.GPXEditorPreferences;
