+++ src/main/java/tf/gpx/edit/algorithms/smoother/IWaypointSmoother.java
@@ -23,88 +23,53 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.algorithms;
+package tf.gpx.edit.algorithms.smoother;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import tf.gpx.edit.helper.GPXEditorPreferences;
 import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.leafletmap.LatLonElev;
 
 /**
- * Wrapper for different smoothing algorithms.
+ * Common interface for all reduce algorithms.
  * 
  * @author thomas
  */
-public class WaypointSmoothing implements IWaypointSmoother {
-    private final static WaypointSmoothing INSTANCE = new WaypointSmoothing();
-
-    public static enum SmoothingAlgorithm {
-        Hampel,
-        SavitzkyGolay,
-        DoubleExponential
-    }
-    
-    private WaypointSmoothing() {
-        super();
-        // Exists only to defeat instantiation.
-    }
-
-    public static WaypointSmoothing getInstance() {
-        return INSTANCE;
-    }
-
+public interface IWaypointSmoother {
     /**
-     * Smoothing a list of values by the requested algorithm.
+     * Entry for list of Doubles.
      * 
-     * @param data list of values
-     * @param smoothingAlgo What SmoothingAlgorithm to use
+     * @param data Double values to process
      * @param dummy Needed to avoid compiler errors due to same signature after erasure as the GPXWaypoint method
-     * @return smoothed list of values
+     * @return List of processed Double values
      */
-    protected static List<Double> apply(
-            final List<Double> data, 
-            final SmoothingAlgorithm smoothingAlgo, 
-            boolean dummy) {
-        List<Double> result = new ArrayList<>(data);
-        
-        switch (smoothingAlgo) {
-            case Hampel:
-                result = HampelSmoother.getInstance().apply(result, dummy);
-                break;
-            case SavitzkyGolay:
-                result = SavitzkyGolaySmoother.getInstance().apply(result, dummy);
-                break;
-            case DoubleExponential:
-                result = DoubleExponentialSmoother.getInstance().apply(result, dummy);
-                break;
-        }
-        
-        return result;
-    }
+    List<Double> apply(final List<Double> data, final boolean dummy);
     
-    public static List<LatLonElev> apply(
-            final List<GPXWaypoint> data, 
-            final SmoothingAlgorithm smoothingAlgo) {
+    /**
+     * Entry for list of GPXWaypoints.
+     * 
+     * lat / lon / elev values are filtered independently and new LatLonElev are built from the results.
+     * 
+     * @param data GPXWaypoints to process
+     * @return List of derived LatLonElev values
+     */
+    default List<LatLonElev> apply(final List<GPXWaypoint> data) {
         // assumption: lat / lon /elevation are independent with respect to fluctuations that we want to eliminate
         // we could apply the algorithm not to the lat / lon values but to the distance/time / course between points calculated from it...
         final List<Double> newLatValues = apply(
                 data.stream().map((t) -> {
                     return t.getLatitude();
-                }).collect(Collectors.toList()), smoothingAlgo, true);
+                }).collect(Collectors.toList()), true);
         final List<Double> newLonValues = apply(
                 data.stream().map((t) -> {
                     return t.getLongitude();
-                }).collect(Collectors.toList()), smoothingAlgo, true);
-        
-        List<Double> newElevValues = 
+                }).collect(Collectors.toList()), true);
+        // not using GPXEditorPreferences.SMOOTHING_ELEVATION here since this is only fpr WaypointSmoothing class...
+        final List<Double> newElevValues = apply(
                 data.stream().map((t) -> {
                     return t.getElevation();
-                }).collect(Collectors.toList());
-        if (GPXEditorPreferences.DO_SMOOTHING_FOR_ELEVATION.getAsType()) {
-            newElevValues = apply(newElevValues, smoothingAlgo, true);
-        }
+                }).collect(Collectors.toList()), true);
         
         final List<LatLonElev> result = new ArrayList<>();
         for (int i = 0; i< data.size(); i++) {
@@ -112,20 +77,5 @@ public class WaypointSmoothing implements IWaypointSmoother {
         }
         
         return result;
-    }
-
-    @Override
-    public List<Double> apply(List<Double> data, boolean dummy) {
-        return apply(
-                data, 
-                GPXEditorPreferences.SMOOTHING_ALGORITHM.getAsType(), 
-                dummy);
-    }
-    
-    @Override
-    public List<LatLonElev> apply(final List<GPXWaypoint> data) {
-        return apply(
-                data, 
-                GPXEditorPreferences.SMOOTHING_ALGORITHM.getAsType());
     }
 }
