+++ src/main/java/tf/gpx/edit/algorithms/smoother/SavitzkyGolaySmoother.java
@@ -23,7 +23,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.algorithms;
+package tf.gpx.edit.algorithms.smoother;
 
 import java.util.Arrays;
 import java.util.List;
