+++ src/main/java/tf/gpx/edit/helper/GPXEditorPreferences.java
@@ -37,8 +37,8 @@ import java.util.prefs.InvalidPreferencesFormatException;
 import java.util.prefs.Preferences;
 import tf.gpx.edit.algorithms.EarthGeometry;
 import tf.gpx.edit.algorithms.WaypointMatching;
-import tf.gpx.edit.algorithms.WaypointReduction;
-import tf.gpx.edit.algorithms.WaypointSmoothing;
+import tf.gpx.edit.algorithms.reducer.WaypointReduction;
+import tf.gpx.edit.algorithms.smoother.WaypointSmoothing;
 import tf.gpx.edit.elevation.ElevationProviderOptions;
 import tf.gpx.edit.elevation.SRTMDataOptions;
 import tf.gpx.edit.elevation.SRTMDownloader;
