+++ src/main/java/tf/gpx/edit/helper/GPXStructureHelper.java
@@ -36,8 +36,8 @@ import javafx.collections.ObservableList;
 import javafx.scene.control.TreeItem;
 import org.apache.commons.collections4.CollectionUtils;
 import tf.gpx.edit.algorithms.WaypointMatching;
-import tf.gpx.edit.algorithms.WaypointReduction;
-import tf.gpx.edit.algorithms.WaypointSmoothing;
+import tf.gpx.edit.algorithms.reducer.WaypointReduction;
+import tf.gpx.edit.algorithms.smoother.WaypointSmoothing;
 import tf.gpx.edit.items.GPXFile;
 import tf.gpx.edit.items.GPXLineItem;
 import tf.gpx.edit.items.GPXLineItemHelper;
