+++ src/main/java/tf/gpx/edit/main/PreferenceEditor.java
@@ -70,8 +70,8 @@ import javafx.util.converter.IntegerStringConverter;
 import jfxtras.styles.jmetro.JMetro;
 import jfxtras.styles.jmetro.Style;
 import tf.gpx.edit.algorithms.EarthGeometry;
-import tf.gpx.edit.algorithms.WaypointReduction;
-import tf.gpx.edit.algorithms.WaypointSmoothing;
+import tf.gpx.edit.algorithms.reducer.WaypointReduction;
+import tf.gpx.edit.algorithms.smoother.WaypointSmoothing;
 import tf.gpx.edit.elevation.ElevationProviderOptions;
 import tf.gpx.edit.elevation.SRTMDataOptions;
 import tf.gpx.edit.elevation.SRTMDownloader;
