+++ src/test/java/tf/gpx/edit/algorithms/TestDoubleExponentialSmoother.java
@@ -25,6 +25,7 @@
  */
 package tf.gpx.edit.algorithms;
 
+import tf.gpx.edit.algorithms.smoother.DoubleExponentialSmoother;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
