+++ src/test/java/tf/gpx/edit/algorithms/TestSmoothing.java
@@ -25,6 +25,7 @@
  */
 package tf.gpx.edit.algorithms;
 
+import tf.gpx.edit.algorithms.smoother.WaypointSmoothing;
 import java.io.File;
 import java.text.DecimalFormatSymbols;
 import java.time.Instant;
